ver 0.4.1
完成 ui 的 callback
還沒能繪製 ui
投影範圍與螢幕座標不同，待修正

+++ nstdlib.h
@@ -62,8 +62,8 @@ namespace nstd {
 		vector<list<dic<T>>> table;
 		int size;//size of table
 
-		int preHashing(string key) {
-			int exp = 8, key_value = 0, p = 1;
+		unsigned long long int preHashing(string key) {
+			unsigned long long int exp = 8, key_value = 0, p = 1;
 			for (int i = key.size(); i > 0; --i) {
 				key_value += key[i] * p;
 				p *= exp;
